clearsym/backend/requirements.txt
alembic==1.18.4
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
flask-cors==6.0.2
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==4.1.0
marshmallow-sqlalchemy==1.4.2
mdurl==0.1.2
Pygments==2.19.2
python-dotenv==1.2.1
requests==2.32.5
rich==14.2.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.3
